Fix ContactPerson XML Generation (#59)
The builder for ContactPerson was attempting to use meta-programming to set 5
sub-fields: company, given_name, sur_name, telephone and email_address.  However,
each of these fields was creating a "Company" node under ContactPerson, like this:

```
<ContactPerson contactType="technical">
  <Company>ACME Corporation</Company>
  <Company>Road</Company>
  <Company>Runner</Company>
  <Company>mailto:acme@example.com</Company>
  <Company>1-800-555-5555</Company>
</ContactPerson>
```

This is not valid according to the SAML spec.  When a program attempted to
validate the metadata XML document, it failed.  For our application, it meant
that occasionally we were getting errors reported.

This commit changes the ContactPerson node to be valid according to the SAML
spec.  After applying this commit, the above invalid SAML will become this:

```
<ContactPerson contactType="technical">
  <Company>ACME Corporation</Company>
  <GivenName>Road</GivenName>
  <SurName>Runner</SurName>
  <EmailAddress>mailto:acme@example.com</EmailAddress>
  <TelephoneNumber>1-800-555-5555</TelephoneNumber>
</ContactPerson
```

Section 2.3.2.2 of the SAML spec (https://docs.oasis-open.org/security/saml/v2.0/saml-metadata-2.0-os.pdf)
has the details on these fields.

I tried a few different ways to set these fields using meta-programming, but
XmlBuilder wasn't liking what I tried.  It looks like using the `tag!` in
builder, you could set tags using the same arguments from method missing, but a
mapping between the fields in metadata_builder and the XML would need to be
provided. Ultimately, I just put in the 5 fields.