Allow configuring NameID format for assertions.
Previously the code would not allow any easy way to set the NameID
format on a per-service-provider or per-request basis. However, this is
a very common pattern because different service providers may expect
different NameID formats (e.g. persistent vs email).

This is difficult to pass through because the NameID generation is
buried in many layers of abstraction. Rather than having to edit all of
the many layers of methods, all of which use positional arguments so
changing them is somewhat fraught, we check for a method
`name_id_format` on the principal object. If the principal responds to
`name_id_format`, we call that method to get the Hash containing
information on the Name ID format and a getter function that we can use
to populate the value of the NameID field.

This matches the earlier pattern of passing asserted_attributes through
a method of that name on the principal object.